Move parse manuscript report functions to manuscript worker class
